server/project_guardian_ai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project-guardian-ai"
version = "0.1.0"
description = "RiskPulse AI — Multi-Agent Risk Monitoring Backend"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.130.0",
    "crewai-tools>=0.45.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
]

[tool.crewai]
type = "crew"

[tool.hatch.build.targets.wheel]
packages = ["src/project_guardian_ai"]
